AI_Code/BackTracking/requirements.txt
opencv-python>=4.5.0
face-recognition>=1.3.0
numpy>=1.19.0
dlib>=19.22.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
tqdm>=4.65.0
python-socketio[client]